remove toolbar sticking/collapsing for now
flip positions of tabbar and toolbar (because of count contexts and stuff)

also revert position of collapse/expand all since toolbar is not pinnable right now
