+++ resources/css/bem/sticky-toolbar.less
@@ -7,8 +7,4 @@
   // Reference: https://bugs.chromium.org/p/chromium/issues/detail?id=1261852
   top: calc(var(--navbar-height) - 1px);
   z-index: @z-index--sticky-toolbar;
-
-  &--with-shadow {
-    .default-box-shadow();
-  }
 }
